+++ example/lib/main.dart
@@ -1,5 +1,4 @@
-import 'package:example/page/grouped_gridview_page.dart';
-import 'package:example/page/grouped_listview_page.dart';
+import 'package:example/page/home_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -13,7 +12,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Grouped Scroll View Demo',
       theme: ThemeData(
         // This is the theme of your application.
         //
@@ -26,66 +25,7 @@ class MyApp extends StatelessWidget {
         // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: const MainAppPage(title: 'Grouped Scroll View Demo'),
-    );
-  }
-}
-
-class MainAppPage extends StatefulWidget {
-  const MainAppPage({Key? key, required this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MainAppPage> createState() => _MainAppState();
-}
-
-class _MainAppState extends State<MainAppPage> {
-  int _curIdx = 0;
-  final List<Widget> _widgets = [
-    const GroupedGridViewTestPage(crossAxisCount: 3),
-    const GroupedListViewTestPage()
-  ];
-  final List<String> _titles = ['GroupedGrid', 'GroupedList'];
-
-  @override
-  Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(_titles[_curIdx]),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: _widgets.elementAt(_curIdx),
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        items: [
-          BottomNavigationBarItem(
-              icon: const Icon(Icons.group), label: _titles.elementAt(0)),
-          BottomNavigationBarItem(
-              icon: const Icon(Icons.group), label: _titles.elementAt(1)),
-        ],
-        currentIndex: _curIdx,
-        onTap: (idx) => setState(() => {_curIdx = idx}),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
+      home: const HomePage(title: 'Grouped Scroll View Demo'),
     );
   }
 }