+++ example/lib/page/grouped_gridview_page.dart
@@ -4,41 +4,55 @@ import 'package:grouped_scroll_view/grouped_scroll_view.dart';
 
 class GroupedGridViewTestPage extends StatefulWidget {
   final int crossAxisCount;
+  final GroupedToggleType? toggleType;
+  final String title;
+  final bool grouped;
 
-  const GroupedGridViewTestPage({super.key, required this.crossAxisCount});
+  const GroupedGridViewTestPage(
+      {super.key, required this.crossAxisCount, this.toggleType, required this.title, this.grouped = true});
 
   @override
-  State<GroupedGridViewTestPage> createState() =>
-      _GroupedGridViewTestPageState();
+  State<GroupedGridViewTestPage> createState() => _GroupedGridViewTestPageState();
 }
 
 class _GroupedGridViewTestPageState extends State<GroupedGridViewTestPage> {
+  GroupedToggleController? _toggleController;
+
   @override
   Widget build(BuildContext context) {
+    if (null != widget.toggleType) {
+      _toggleController = GroupedToggleController(
+        toggleType: widget.toggleType!,
+        onToggleChanged: (int idx, bool selected) =>
+            print('GroupedGridViewTestPage:onToggleChanged===>idx:[$idx]--selected:[$selected]'),
+      );
+    }
     return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
       body: GroupedScrollView.grid(
         gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-            mainAxisSpacing: 5,
-            crossAxisSpacing: 5,
-            crossAxisCount: widget.crossAxisCount),
-        itemGrouper: (Person person) {
-          return person.birthYear;
-        },
-        stickyHeaderBuilder: (BuildContext context, int year, int idx) =>
-            Container(
-          color: Colors.white,
-          padding: const EdgeInsets.all(8),
-          constraints: const BoxConstraints.tightFor(height: 30),
-          child: Text(
-            '$year',
-            style: const TextStyle(fontWeight: FontWeight.bold),
-          ),
-        ),
+            mainAxisSpacing: 5, crossAxisSpacing: 5, crossAxisCount: widget.crossAxisCount),
+        groupedOptions: widget.grouped
+            ? GroupedScrollViewOptions(
+                itemGrouper: (Person person) {
+                  return person.birthYear;
+                },
+                stickyHeaderBuilder: (BuildContext context, int year, int idx) => Container(
+                      color: Colors.white,
+                      padding: const EdgeInsets.all(8),
+                      constraints: const BoxConstraints.tightFor(height: 30),
+                      child: Text(
+                        '$year',
+                        style: const TextStyle(fontWeight: FontWeight.bold),
+                      ),
+                    ))
+            : null,
         itemBuilder: (BuildContext context, Person item) {
           return Container(
             color: Colors.lightGreen,
             padding: const EdgeInsets.all(8),
-            constraints: const BoxConstraints.tightFor(width: 100),
             child: Center(
               child: Text(
                 item.name,
@@ -80,7 +94,14 @@ class _GroupedGridViewTestPageState extends State<GroupedGridViewTestPage> {
             ),
           ],
         ),
+        toggleController: _toggleController,
       ),
     );
+  }
+
+  @override
+  void dispose() {
+    _toggleController?.dispose();
+    super.dispose();
   }
 }