+++ lib/grouped_scroll_view.dart
@@ -1,3 +1,7 @@
 library grouped_scroll_view;
 
 export 'src/grouped_scroll_view.dart';
+export 'src/grouped_scroll_view_options.dart';
+export 'src/toggle/toggle_style.dart';
+export 'src/toggle/toggle_controller.dart';
+export 'src/toggle/toggle_type.dart';