+++ lib/src/toggle/toggle_type.dart
@@ -1,0 +1,5 @@
+/// event types
+enum GroupedToggleType {
+  single,
+  multiple,
+}